backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
faiss-cpu>=1.12.0
openai==1.82.0
python-dotenv==1.1.0
numpy
